requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
opencv-python>=4.5.0
astropy>=4.2
matplotlib>=3.4.0
pyyaml>=5.4.0
pyevtk>=1.4.0
pandas>=1.3.0
pytest>=6.2.0
pytest-cov>=2.12.0
sphinx>=4.0.0
black>=21.5b2
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910